♻️ Refactor ReferralDialog into reusable component
